Rename visualizer module to visualize
